+++ projects/engine/core/public/core/diagnostic/source_location/SourceLocation.hpp
@@ -38,8 +38,8 @@ public:
 
 	////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 	static inline consteval SourceLocation					Current(
-		size_t												line			= __builtin_LINE(),
-		size_t												column			= __builtin_COLUMN(),
+		u64													line			= __builtin_LINE(),
+		u64													column			= __builtin_COLUMN(),
 		const char										*	file			= __builtin_FILE(),
 		const char										*	function		= __builtin_FUNCTION()
 	) noexcept
@@ -53,13 +53,13 @@ public:
 	}
 
 	////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-	inline constexpr size_t 								GetLine() const noexcept
+	inline constexpr u64 									GetLine() const noexcept
 	{
 		return line;
 	}
 	
 	////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-	inline constexpr size_t 								GetColumn() const noexcept
+	inline constexpr u64	 								GetColumn() const noexcept
 	{
 		return column;
 	}
@@ -79,8 +79,8 @@ public:
 private:
 
 	////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-	size_t													line			= {};
-	size_t													column			= {};
+	u64														line			= {};
+	u64														column			= {};
 	const char											*	file			= {};
 	const char											*	function		= {};
 };
